Fix the build on Windows
Looks like Ruby on Windows has fewer bits for numbers, so f64 or i64 may
not fit in a Fixnum. Magnus returns these larger numbers in the as Error
component of the Result. From the perspective of sending it back to
Ruby, it's fair game, so we can unwrap the error.