Implement `Caller#export`
Implement `Caller#export` which allows getting a instance's memories
or funcs from a `Caller`.

This is a a little odd because the caller itself only "lives" for the
duration of call, but we can't control the lifetime of Ruby objects
we hand out:

```ruby
Func.new(...) do |caller|
  $yes_i_can_do_this = caller
end
```

The approach I'm taking here is to model a Caller's "context" as an
`Option`. While the func is being called, it'll be `Some`, but we'll
change it to `None` before exiting the func call.

Thus extracting the context from a `Caller` can fail. That's
implemented with an `Error` that gets surfaced as  Ruby exception.
See the added test.