Add support for component enum type
Decision: the enum string is not validated against the list of possible
values. The rationale is that Wasmtime will validate it at a later
point, thus by adding an extra validation we're incurring the cost
twice.

The tradeoff is that we need to rely on Wasmtime's error message which
does not include the list of valid values.

An implementation of validation would look like:
```rust
// SAFETY: we're only calling into Ruby after we're done using the str
let rstring_str = unsafe { rstring.as_str()? };
if !enum_.names().any(|name| name == rstring_str) {
    return err!(
        "invalid enum value \"{}\", valid values: {:?}",
        rstring_str,
        RArray::from_iter(enum_.names())
    );
}
```

And the test:
```ruby
["enum", "no", /invalid enum value "no", valid values: \["s", "m", "l"\]/],
```